--- card/ 测试分类|商品1.txt
@@ -1,6 +1,0 @@
-hcjkasnckas
-Hbsackcsjaa
-Qhruiwqlkvv
-Iuehowqncav
-Qwojfosamsv
-wiqjvsmavds
