+++ main.py
@@ -1,5 +1,5 @@
 from func import run_bot
-from func import check_trade
+from user import check_trade
 import threading
 
 
@@ -7,9 +7,5 @@ thread = threading.Thread(target=check_trade)
 thread.start()
 
 run_bot()
-
-
-
-
