+++ requirements.txt
@@ -6,6 +6,9 @@ decorator==4.4.2
 future==0.18.2
 idna==2.9
 pycparser==2.20
+pycryptodomex==3.9.4
+pyOpenSSL==19.1.0
+python-alipay-sdk==2.0.1
 python-telegram-bot==12.7
 requests==2.23.0
 six==1.15.0
